Final Release-Candidate corrections

+++ BugTracker/tracker/BugTracker/mainwindow.hpp
@@ -60,7 +60,7 @@ private:
     void load_query_intoMemory(QString command);
     void load_tree_fromMemory();
     bool load_settings(std::vector<QString>& dbparams, int& port);
-    std::vector<QString> edit_memoryItem(int item_position, int & code);
+    std::vector<QString> edit_memoryItem(size_t item_position, int & code);
     void add_new_item();
     QString sqlInsert_fromValues(QStringList values);
 };
